kiểm tra khia delete bill,brand và thông báo lỗi khi add bill, product

+++ src/main/java/com/example/controllers/ComponentControllers/Form/BrandFormController.java
@@ -3,6 +3,7 @@ package com.example.controllers.ComponentControllers.Form;
 import com.example.models.Brand;
 import com.example.services.BrandService;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
 import javafx.scene.control.TextField;
 import javafx.stage.Stage;
@@ -71,9 +72,18 @@ public class BrandFormController {
   }
   
   public void handleBtnDelete() {
-    BrandService brandService = new BrandService();
-    brandService.deleteBrand(this.brand.getBrandId());
-    closeForm();
+      BrandService brandService = null;
+      if (brandService.hasProduct(this.brand.getBrandId())) {
+          Alert alert = new Alert(Alert.AlertType.WARNING);
+          alert.setTitle("Cannot Delete Brand");
+          alert.setHeaderText(null);
+          alert.setContentText("Brand không thể xóa vì có sản phẩm liên quan.");
+          alert.showAndWait();
+          return;
+      }
+      brandService = new BrandService();
+      brandService.deleteBrand(this.brand.getBrandId());
+      closeForm();
   }
   
   public void handleBtnCancel() {
